refactor(issue-95): split oversized config and protocol files
