+++ tests/cmake/setup.cmake
@@ -3,6 +3,8 @@ if(NOT DEFINED GRAVITY_RUNTIME_PROTOCOL_SOURCES)
         "${GRAVITY_ROOT_DIR}/runtime/src/protocol/BackendJsonCodec.cpp"
         "${GRAVITY_ROOT_DIR}/runtime/src/protocol/BackendJsonCodecParse.cpp"
         "${GRAVITY_ROOT_DIR}/runtime/src/protocol/BackendJsonCodecReadNumber.cpp"
+        "${GRAVITY_ROOT_DIR}/runtime/src/protocol/BackendJsonCodecParseStatus.cpp"
+        "${GRAVITY_ROOT_DIR}/runtime/src/protocol/BackendJsonCodecParseSnapshot.cpp"
         "${GRAVITY_ROOT_DIR}/runtime/src/protocol/BackendClient.cpp"
         "${GRAVITY_ROOT_DIR}/runtime/src/protocol/BackendProtocol.cpp"
     )
