+++ tests/cmake/targets_gtests.cmake
@@ -11,6 +11,8 @@ if(GRAVITY_TEST_UNIT_CONFIG_SOURCES)
             "${GRAVITY_ROOT_DIR}/engine/src/config/SimulationArgsInitStateOptions.cpp"
             "${GRAVITY_ROOT_DIR}/engine/src/config/SimulationArgsFluidOptions.cpp"
             "${GRAVITY_ROOT_DIR}/engine/src/config/SimulationOptionRegistry.cpp"
+            "${GRAVITY_ROOT_DIR}/engine/src/config/SimulationOptionRegistryApply.cpp"
+            "${GRAVITY_ROOT_DIR}/engine/src/config/SimulationOptionRegistryEntries.cpp"
             "${GRAVITY_ROOT_DIR}/engine/src/config/SimulationConfig.cpp"
             "${GRAVITY_ROOT_DIR}/engine/src/config/SimulationModes.cpp"
             "${GRAVITY_ROOT_DIR}/engine/src/config/TextParse.cpp"
@@ -25,6 +27,8 @@ if(GRAVITY_TEST_UNIT_PROTOCOL_SOURCES)
             "${GRAVITY_ROOT_DIR}/runtime/src/protocol/BackendJsonCodec.cpp"
             "${GRAVITY_ROOT_DIR}/runtime/src/protocol/BackendJsonCodecParse.cpp"
             "${GRAVITY_ROOT_DIR}/runtime/src/protocol/BackendJsonCodecReadNumber.cpp"
+            "${GRAVITY_ROOT_DIR}/runtime/src/protocol/BackendJsonCodecParseStatus.cpp"
+            "${GRAVITY_ROOT_DIR}/runtime/src/protocol/BackendJsonCodecParseSnapshot.cpp"
             "${GRAVITY_ROOT_DIR}/runtime/src/protocol/BackendProtocol.cpp"
             "${GRAVITY_ROOT_DIR}/engine/src/config/TextParse.cpp"
     )
@@ -87,6 +91,8 @@ set(GRAVITY_TEST_BASE_RUNTIME_SOURCES
     "${GRAVITY_ROOT_DIR}/runtime/src/frontend/FrontendRuntime.cpp"
     "${GRAVITY_ROOT_DIR}/runtime/src/frontend/FrontendCommon.cpp"
     "${GRAVITY_ROOT_DIR}/engine/src/config/SimulationOptionRegistry.cpp"
+    "${GRAVITY_ROOT_DIR}/engine/src/config/SimulationOptionRegistryApply.cpp"
+    "${GRAVITY_ROOT_DIR}/engine/src/config/SimulationOptionRegistryEntries.cpp"
     "${GRAVITY_ROOT_DIR}/engine/src/config/SimulationConfig.cpp"
     "${GRAVITY_ROOT_DIR}/engine/src/config/SimulationModes.cpp"
     ${GRAVITY_TEST_BASE_BRIDGE_SOURCES}
